pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "book-normalizer"
version = "0.1.0"
description = "Semi-automatic Russian book source normalizer for TTS pipelines"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "books-to-audio team" },
]

dependencies = [
    "pydantic>=2.5,<3",
    "click>=8.1,<9",
    "PyMuPDF>=1.24,<2",
    "python-docx>=1.1,<2",
    "ebooklib>=0.18,<1",
    "lxml>=5.0,<6",
    "rich>=13.0,<14",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9",
    "pytest-cov>=5.0,<6",
    "ruff>=0.4,<1",
]

[project.scripts]
normalize-book = "book_normalizer.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/book_normalizer"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
